third_party/WebKit/LayoutTests/fast/scroll-behavior/wheel-and-touch-scroll-use-count.html

<!DOCTYPE HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>

  ::-webkit-scrollbar {
    display: none;
  }
  body {
    margin: 0px;
    height: 1000px;
    width: 1000px;
  }
  #scrollable {
    background-color: #FF7F7F;
    height: 600px;
    width: 600px;
    overflow: scroll;
  }
  #content {
    height: 700px;
    width: 700px;
  }

</style>

<div id="scrollable">
	<div id="content"></div>
</div>

<script>

var scrollable = document.getElementById('scrollable');
var rect = scrollable.getBoundingClientRect();
var start_x = (rect.left + rect.right) / 2;
var start_y = (rect.top + rect.bottom) / 2;

function scrollDown(pixels_to_scroll, start_x, start_y,
    gesture_source_type, speed_in_pixels_s) {
  return new Promise((resolve, reject) => {
    chrome.gpuBenchmarking.smoothScrollBy(pixels_to_scroll,
                                          resolve,
                                          start_x,
                                          start_y,
                                          gesture_source_type,
                                          'down',
                                          speed_in_pixels_s);
  });
}

const MAX_RAF = 1000;
function waitForScrollEnd() {
  var last_scroll_offset = scrollable.scrollTop;
  var last_changed_count = 0;
  return new Promise((resolve, reject) => {
    function tick(raf_count) {
    // We requestAnimationFrame either for 1000 frames or until 20 frames with
    // no change have been observed.
      if (raf_count >= MAX_RAF || raf_count - last_changed_count > 20) {
        resolve();
      } else {
        if (scrollable.scrollTop != last_scroll_offset) {
          last_changed_count = raf_count;
          last_scroll_offset = scrollable.scrollTop;
        }
        requestAnimationFrame(tick.bind(this, raf_count + 1));
      }
    }
    tick(0);
  });
}

const TOUCH_SOURCE_TYPE = 1; // TOUCH_INPUT from synthetic_gesture_params.h
const WHEEL_SOURCE_TYPE = 2; // MOUSE_INPUT from synthetic_gesture_params.h

var SCROLL_BY_TOUCH_COUNTER = 1847; // from enum Feature in UseCounter.h
var SCROLL_BY_WHEEL_COUNTER = 1848; // from enum Feature in UseCounter.h

promise_test(t => {
  return scrollDown(50, start_x, start_y , WHEEL_SOURCE_TYPE, 4000)
  .then(waitForScrollEnd)
  .then(() => {
    assert_true(internals.isUseCounted(document, SCROLL_BY_WHEEL_COUNTER));
  });
}, "Scrolling by wheel on main/compositor should update usecounters.");

const IS_MAC = navigator.platform.indexOf('Mac') == 0;
if (!IS_MAC) {
  promise_test(t => {
    return scrollDown(50, start_x, start_y , TOUCH_SOURCE_TYPE, 4000)
    .then(waitForScrollEnd)
    .then(() => {
      assert_true(internals.isUseCounted(document, SCROLL_BY_TOUCH_COUNTER));
    });
  }, "Scrolling by touch on main/compositor should update usecounters.");
}

</script>
